Add ticket editing and creation pages
