+++ client/src/App.tsx
@@ -21,7 +21,8 @@ import PrerequisAdminPage   from './pages/PrerequisAdminPage';
 import ManagerDashboardPage from './pages/ManagerDashboardPage';
 import RegisterUserPage     from './pages/RegisterUserPage';
 import NotificationsPage    from './pages/NotificationsPage';
-import TicketsPage          from './pages/TicketsPage';
+import TicketsListPage      from './pages/TicketsListPage';
+import CreateTicketPage     from './pages/CreateTicketPage';
 
 import Footer from './components/Footer';
 
@@ -52,7 +53,8 @@ function RoleRoutes() {
         <Route path="/manager/create"              element={<RegisterUserPage />} />
         <Route path="/manager/modules"             element={<AdminModulesPage />} />
         <Route path="/manager/modules/:moduleId"   element={<AdminModuleEditor />} />
-        <Route path="/manager/tickets"            element={<TicketsPage />} />
+        <Route path="/manager/tickets"            element={<TicketsListPage />} />
+        <Route path="/manager/tickets/new"        element={<CreateTicketPage />} />
         <Route path="/admin/*"                     element={<Navigate to="/manager" replace />} />
         <Route path="*"                            element={<Navigate to="/manager" replace />} />
       </Routes>
@@ -68,7 +70,7 @@ function RoleRoutes() {
         <Route path="/admin/modules/:moduleId"     element={<AdminModuleEditor />} />
         <Route path="/admin/prerequis"             element={<PrerequisAdminPage />} />
         <Route path="/admin/notifications"         element={<NotificationsPage />} />
-        <Route path="/admin/tickets"               element={<TicketsPage />} />
+        <Route path="/admin/tickets"               element={<TicketsListPage />} />
         <Route path="/admin/create"                element={<RegisterUserPage />} />
 
         <Route path="/"   element={<Navigate to="/admin" replace />} />
@@ -83,7 +85,8 @@ function RoleRoutes() {
       <Route path="/"                   element={<HomePage />} />
       <Route path="/modules/:moduleId"  element={<ModulePage />} />
       <Route path="/favoris"            element={<FavoritesPage />} />
-      <Route path="/tickets"            element={<TicketsPage />} />
+      <Route path="/tickets"            element={<TicketsListPage />} />
+      <Route path="/tickets/new"        element={<CreateTicketPage />} />
       <Route path="*"                   element={<Navigate to="/" replace />} />
     </Routes>
   );
